+++ client/src/components/ServiceCard/ServiceCard.css
@@ -1,10 +1,4 @@
-.service-card {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  margin: 8px;
-}
+
 
 .service-card .service-category {
   color: white;
@@ -20,15 +14,21 @@
   font-family: "MarcellusSC", serif;
 }
 
-.service-card .service-button {
+.service-card .service-category {
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: column;
+
+}
+
+.service-card .service-button {
+ 
   background-color: #8fa4b1;
   border-radius: 32px;
   color: white;
-  max-width: 300px;
+
+  
 }
 
 .service-card .image-service {
@@ -37,7 +37,7 @@
 }
 
 .service-card .members {
-  margin: 16px;
+  /* gap: 16px; */
   padding: 8px;
 }
 
@@ -48,14 +48,15 @@
   flex-direction: column;
   background-color: #8fa4b1;
   border-radius: 32px;
-  margin: 16px;
-  max-width: 200px;
+  margin: 24px;
+  max-width: 100%;
 }
 
 .service-card .image-member {
-  width: 20vh;
+  width: 25vh;
   margin: 8px;
 }
+
 
 @media (min-width: 800px) {
   .service-card .service-button {
@@ -75,6 +76,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    margin: 32px;
+    
   }
 }
