+++ client/src/pages/AboutPage/AboutPage.css
@@ -1,19 +1,43 @@
+
+
 .about-main {
   background-color: #18242b;
-  margin: 0;
-  padding: 0;
+  background-image: url("aboutbackground/curve-line-top.svg"), url("aboutbackground/curve-line-bottom.svg");
+  background-size: 200% auto, 200% auto;
+  background-position: top , bottom;
+  background-repeat: no-repeat, no-repeat;
 }
 
 .about-main .about-content {
   display: flex;
   flex-direction: column;
   text-align: center;
-  padding: 2rem;
+  padding: 32px;
   color: white;
+  
+}
+
+.about-main .about-history {
+  margin-bottom: 32px;
+}
+
+.about-main .about-history h2 {
+  margin: 48px;
+  padding: 16px;
+  
+  
+  
+}
+
+.about-main .about-history p {
+  font-size: 1.2rem;
+  
 }
 
 .about-main .about-history h1 {
   margin-top: 3rem;
+  
+  
 }
 
 .services-container {
@@ -21,10 +45,13 @@
   flex-direction: column;
   align-items: center;
   flex-direction: column;
+  gap: 48px;
 }
 
-@media screen and (min-width: 768px) {
+@media screen and (min-width: 800px) {
   .about-main {
     min-height: calc(100vh - 134.34px);
+    
+    
   }
 }
